The <span class="wbr"> hack seems not worth it, removed.
In principle it makes the zwsp render as a normal space rather than some
unfortunate glyph in browsers that handle zwsp badly, but in practice those
browsers also don't seem to support all that much css (except semi-new
versions of opera, that render a glyph but do support hiding it). But it
stops the breaking behaviour of the zwsp in webkit, so it's clearly not
worth it.