Try to make image rotation work in more browsers.
I use Opera. The old version worked in Opera. Looking at the standards, it
should work in everything with SVG support. So of course it worked in
nothing except Opera. Using <object> makes me a little sad, but it does
allow for showing a message and unrotated image in browsers without SVG
support. Plus it works in most browsers, that's a plus too.