Use CSS instead of an SVG to rotate the image.
All of this to because some browsers use the EXIF rotation in the image,
and I can turn that off for <img> (with image-orientation: none), but I
can't inside the SVG.

Plus browsers without SVG support can now at least show the unrotated
image.

Too bad it's so gross. I use CSS variables, which is of course not
particularly compatible, but I tried to make sure it breaks in reasonable
ways when unsupported.